Edit transaction activity added

+++ app/src/main/java/com/franksacco/wallet/SettingsActivity.java
@@ -5,6 +5,7 @@ import android.app.Dialog;
 import android.app.DialogFragment;
 import android.content.Context;
 import android.content.DialogInterface;
+import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Build;
 import android.os.Bundle;
@@ -12,7 +13,6 @@ import android.preference.ListPreference;
 import android.preference.Preference;
 import android.support.annotation.NonNull;
 import android.support.design.widget.Snackbar;
-import android.support.v4.app.NavUtils;
 import android.support.v7.app.ActionBar;
 import android.preference.PreferenceFragment;
 import android.preference.PreferenceManager;
@@ -21,7 +21,10 @@ import android.util.Log;
 import android.view.MenuItem;
 
 import com.franksacco.wallet.helpers.CurrencyManager;
+import com.franksacco.wallet.helpers.CurrencyUpdater;
 import com.franksacco.wallet.helpers.DatabaseOpenHelper;
+
+import java.util.ArrayList;
 
 
 /**
@@ -57,7 +60,6 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
      */
     private static void bindPreferenceChange(Preference preference) {
         preference.setOnPreferenceChangeListener(changeListener);
-
         changeListener.onPreferenceChange(preference,
                 PreferenceManager
                         .getDefaultSharedPreferences(preference.getContext())
@@ -109,7 +111,7 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
         public boolean onOptionsItemSelected(MenuItem item) {
             int id = item.getItemId();
             if (id == android.R.id.home) {
-                NavUtils.navigateUpFromSameTask(getActivity());
+                this.getActivity().onBackPressed();
                 return true;
             }
             return super.onOptionsItemSelected(item);
@@ -122,10 +124,11 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
             switch (key) {
                 case "settings_currency_update":
                     AppCompatPreferenceActivity activity =
-                            (AppCompatPreferenceActivity) getActivity();
-                    Snackbar.make(activity.getListView(),
-                            "TODO: download currency rates", Snackbar.LENGTH_SHORT)
-                            .show();
+                            (AppCompatPreferenceActivity) this.getActivity();
+
+                    // todo fix
+                    new CurrencyUpdater(activity).execute();
+
                     return true;
                 case "settings_data_deleteAll":
                     DialogFragment dialog = new ConfirmDeleteDialog();
@@ -146,27 +149,29 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
                     .setIcon(R.drawable.ic_delete_forever_black_24dp)
                     .setPositiveButton(android.R.string.ok, new DialogInterface.OnClickListener() {
                         public void onClick(DialogInterface dialog, int id) {
-                            deleteAllData();
+                            ConfirmDeleteDialog.this.deleteAllData();
                         }
                     })
                     .setNegativeButton(android.R.string.cancel, null);
             return builder.create();
         }
 
+        /**
+         * Delete database, preferences and restart application
+         */
         private void deleteAllData() {
-            AppCompatPreferenceActivity activity = (AppCompatPreferenceActivity) getActivity();
-            String message = activity.getResources().getString(R.string.deleteAll_ok);
-            try {
-                if (!activity.deleteDatabase(DatabaseOpenHelper.DATABASE_NAME)) {
-                    throw new Error("Error during database deletion");
-                }
-                PreferenceManager.getDefaultSharedPreferences(activity.getBaseContext())
-                        .edit().clear().apply();
-            } catch (Error e) {
-                message = e.getMessage();
+            AppCompatPreferenceActivity activity = (AppCompatPreferenceActivity) this.getActivity();
+
+            activity.deleteDatabase(DatabaseOpenHelper.DATABASE_NAME);
+            PreferenceManager.getDefaultSharedPreferences(activity.getBaseContext())
+                    .edit().clear().apply();
+
+            Intent i = activity.getBaseContext().getPackageManager()
+                    .getLaunchIntentForPackage(activity.getBaseContext().getPackageName());
+            if (i != null) {
+                i.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
             }
-            Snackbar.make(activity.getListView(), message, Snackbar.LENGTH_SHORT)
-                    .show();
+            this.startActivity(i);
         }
     }
